requirements.txt
absl-py==0.1.10
aiml==0.8.6
alabaster==0.7.10
anaconda-client==1.6.5
anaconda-navigator==1.6.9
anaconda-project==0.8.0
appdirs==1.4.3
asn1crypto==0.22.0
astroid==1.5.3
astropy==2.0.2
babel==2.5.0
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==1.5.0
bokeh==0.12.10
boto==2.48.0
Bottleneck==1.2.1
bz2file==0.98
CacheControl==0.12.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
chatterbot-corpus==1.1.1
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
colorama==0.3.9
commonregex==1.5.4
comtypes==1.1.2
conda==4.3.29
conda-build==3.0.27
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.0.3
cycler==0.10.0
cymem==1.31.2
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.3
datashape==0.5.4
decorator==4.1.2
dill==0.2.6
Distance==0.1.3
distlib==0.2.5
distributed==1.19.1
Django==1.11.7
django-cors-headers==2.1.0
django-json-field==0.5.7
djangorestframework==3.7.3
docutils==0.14
docx==0.2.4
docx2txt==0.7
en-core-web-sm==1.2.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==2.0.12
Flask==0.12.2
Flask-Cors==3.0.3
ftfy==4.4.2
future==0.16.0
futures==3.1.1
fuzzywuzzy==0.16.0
gensim==3.1.0
gevent==1.2.2
glob2==0.5
google==1.9.3
goslate==1.5.1
greenlet==0.4.12
gTTS==1.2.2
gTTS-token==1.1.1
h5py==2.7.0
heapdict==1.0.0
html5lib==0.9999999
idna==2.6
image==1.5.16
imageio==2.2.0
imagesize==0.7.1
inflect==0.2.5
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.6
jsonfield==2.0.2
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
Keras==2.1.3
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
lockfile==0.12.2
lxml==4.1.0
Markdown==2.6.11
MarkupSafe==1.0
mathparse==0.1.1
matplotlib==2.1.0
mccabe==0.6.1
menuinst==1.4.10
mistune==0.7.4
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
murmurhash==0.26.4
nameparser==0.5.6
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numba==0.35.0
numexpr==2.6.2
numpy==1.13.3
numpydoc==0.7.0
oauthlib==2.0.6
odo==0.5.1
olefile==0.44
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
partd==0.3.8
path.py==10.3.1
pathlib==1.0.1
pathlib2==2.3.0
patsy==0.4.1
pdfminer.six==20170720
pep8==1.7.0
pickleshare==0.7.4
Pillow==4.2.1
pkginfo==1.4.1
plac==0.9.6
ply==3.10
preshed==1.0.0
progress==1.3
prompt-toolkit==1.0.15
protobuf==3.5.1
psutil==5.4.0
py==1.4.34
PyAudio==0.2.11
pycodestyle==2.3.1
pycorenlp==0.3.0
pycosat==0.6.2
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.4.7
pycurl==7.43.0
PyDictionary==1.5.2
pydub==0.20.0
pyflakes==1.6.0
Pygments==2.2.0
pygsr==0.1
pyjarowinkler==1.8
pylint==1.7.4
pymongo==3.6.0
pyodbc==4.0.17
pyOpenSSL==17.2.0
pyparsing==2.2.0
PyPDF2==1.26.0
pypiwin32==220
pyreadline==2.1
PySocks==1.6.7
pytesseract==0.1.7
pytest==3.2.1
python-aiml==0.9.1
python-dateutil==2.6.1
python-docx==0.8.6
python-pdfbox==0.1.0
python-twitter==3.3
pyttsx==1.1
pyttsx3==2.6
pytz==2017.2
PyWavelets==0.5.2
pywin32==221
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
quantulum==0.1.13
rake-nltk==1.0.2
regex==2017.9.23
requests==2.18.4
requests-oauthlib==0.8.0
rope==0.10.5
ruamel-yaml==0.11.14
sarge==0.1.4
scikit-image==0.13.0
scikit-learn==0.19.1
scipy==0.19.1
seaborn==0.8
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
smart-open==1.5.3
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
spacy==1.9.0
speech==0.5.2
SpeechRecognition==3.7.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.2.4
sqlalchemy==1.1.13
statsmodels==0.8.0
stemming==1.0.1
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
tensorflow==1.5.0
tensorflow-tensorboard==1.5.0
termcolor==1.1.0
testpath==0.3.1
textblob==0.13.0
tflearn==0.3.2
Theano==1.0.0rc1
thinc==6.5.2
tika==1.15
toolz==0.8.2
tornado==4.5.2
tqdm==4.19.4
traitlets==4.3.2
typing==3.6.2
ujson==1.35
unicodecsv==0.14.1
urllib3==1.22
Wand==0.4.4
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
widgetsnbextension==3.0.2
wikipedia==1.4.0
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwings==0.11.4
xlwt==1.3.0
zict==0.1.3
